% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar_base.R
\name{atualizar_base_resultados}
\alias{atualizar_base_resultados}
\title{Atualizar Base de Resultados}
\usage{
atualizar_base_resultados(produto = "megasena")
}
\arguments{
\item{produto}{Nome do jogo das Loterias que deseja atualizar a base. Caso não
informado, por padrão será a Megasena \link{megasena}.
Produtos disponíveis:
\itemize{
\item Megasena -> digite 'megasena' \link{megasena}
\item Lotofácil -> digite 'lotofacil' \link{lotofacil}
\item Quina -> digite 'quina' \link{quina}
\item Lotomania -> digite 'lotomania' \link{lotomania}
\item Timemania -> digite 'timemania' \link{timemania}
\item Super Sete -> digite 'supersete' \link{supersete}
\item Dia de Sorte -> digite 'diadesorte' \link{diadesorte}
}}
}
\value{
Uma \code{tibble} - a quantidade de variáveis depende do produto.
}
\description{
Esta função deve ser utilizada para atualizar a base local de resultados.
O pacote atualiza diariamente a base de resultados, portanto, que faz a
instalação do mesmo, sempre recebe uma base atualizada, porém, após instalá-lo,
é necessário que faça a atualização da base sempre que julgar necessário.
}
