% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diadesorte}
\alias{diadesorte}
\title{Dia de Sorte}
\format{
Uma tibble contendo 9 variávies.
\itemize{
\item \strong{data_apuracao} Data do sorteio (character)
\item \strong{concurso} Numero do concurso (numeric)
\item \strong{numeros_sorteados} Todos os números sorteados, separados por ponto
e vírgula (character)
\item \strong{num_1} Primeira dezena do sorteio (numeric)
\item \strong{num_2} Segunda dezena do sorteio (numeric)
\item \strong{num_3} Terceira dezena do sorteio (numeric)
\item \strong{num_4} Quarta dezena do sorteio (numeric)
\item \strong{num_5} Quinta dezena do sorteio (numeric)
\item \strong{num_6} Sexta dezena do sorteio (numeric)
\item \strong{num_7} Sétima dezena do sorteio (numeric)
}
}
\source{
Site da Loteria Federal \url{https://loterias.caixa.gov.br/Paginas/Dia-de-Sorte.aspx}
}
\usage{
diadesorte
}
\description{
Conjunto de dados contendo todos os resultados do produto das Loterias
chamado Dia de Sorte.
}
\note{
Números em ordem crescente e não na ordem em que foram sorteados.
}
\keyword{datasets}
