% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historico_resumido.R
\name{historico_resumido_produtos}
\alias{historico_resumido_produtos}
\title{Resumo Histórico de Todos os Produtos}
\usage{
historico_resumido_produtos()
}
\value{
Uma \code{tibble} contendo sete variáveis.
}
\description{
Utilize esta função para obter um resumo histórico de todos os produtos
das Loterias do Brasil.
}
\details{
Dados presentes no resumo:
\itemize{
\item \strong{nome_produto} (character)
\item \strong{data_primeiro_concurso} (date)
\item \strong{quantidade_concursos} (numeric)
\item \strong{quantidade_concursos_com_ganhador} (numeric)
\item \strong{percentual_com_ganhador} (numeric)
\item \strong{media_premiacao} (numeric)
\item \strong{maior_premio} (numeric)
\item \strong{menor_premio} (numeric)
\item \strong{total_dezenas_sorteadas} (numeric)
\item \strong{numero_mais_sorteado} (numeric)
\item \strong{numero_menos_sorteado} (numeric)
}
}
\examples{
# Gerando o resumo e filtrando o nome do produto, a quatidade de concursos
# e o percentual do concursos em que houve ganhador
historico_resumido_produtos() |>
 dplyr::select(nome_produto, quantidade_concursos, percentual_com_ganhador)

}
