% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotomania}
\alias{lotomania}
\title{Lotomania}
\format{
Uma tibble contendo 22 variaveis.
\itemize{
\item \strong{data_apuracao} Data do sorteio (date)
\item \strong{concurso} Numero do concurso (numeric)
\item \strong{houve_ganhador} Flag indicando se houve ganhador do prêmio máximo no
concurso  <0 ou 1> (numeric)
\item \strong{valor_premio_maximo} Valor total do prêmio máximo do concurso, caso
exista ganhador (numeric)
\item \strong{numeros_sorteados} Todos os números sorteados, separados por ponto
e vírgula (character)
\item \strong{num_1} Primeira dezena do sorteio (numeric)
\item \strong{num_2} Segunda dezena do sorteio (numeric)
\item \strong{num_3} Terceira dezena do sorteio (numeric)
\item \strong{num_4} Quarta dezena do sorteio (numeric)
\item \strong{num_5} Quinta dezena do sorteio (numeric)
\item \strong{num_6} Sexta dezena do sorteio (numeric)
\item \strong{num_7} Sétima dezena do sorteio (numeric)
\item \strong{num_8} Oitava dezena do sorteio (numeric)
\item \strong{num_9} Nona dezena do sorteio (numeric)
\item \strong{num_10} Décima dezena do sorteio (numeric)
\item \strong{num_11} Décima primeira dezena do sorteio (numeric)
\item \strong{num_12} Décima segunda dezena do sorteio (numeric)
\item \strong{num_13} Décima terceira dezena do sorteio (numeric)
\item \strong{num_14} Décima quarta dezena do sorteio (numeric)
\item \strong{num_15} Décima quinta dezena do sorteio (numeric)
\item \strong{num_16} Décima sexta dezena do sorteio (numeric)
\item \strong{num_17} Décima sétima dezena do sorteio (numeric)
\item \strong{num_18} Décima oitava dezena do sorteio (numeric)
\item \strong{num_19} Décima nona dezena do sorteio (numeric)
\item \strong{num_20} Vigésima dezena do sorteio (numeric)
}
}
\source{
Site da Loteria Federal \url{https://loterias.caixa.gov.br/Paginas/Lotomania.aspx}
}
\usage{
lotomania
}
\description{
Conjunto de dados contendo todos os resultados do produto das Loterias
chamado Lotomania.
}
\note{
Números em ordem crescente e não na ordem em que foram sorteados.
}
\keyword{datasets}
