% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produtos_existentes.R
\name{produtos_existentes}
\alias{produtos_existentes}
\title{Lista Produtos Disponíveis}
\usage{
produtos_existentes()
}
\value{
Uma \code{tibble} com três variáveis.
}
\description{
Utilize esta função para receber um conjunto de dados contendo todos os
produtos disponíveis.
}
\details{
Dataset retornado contém as seguintes variáveis:
\itemize{
\item \strong{nome_produto} Nome oficial do produto (character)
\item \strong{parametro_produto} String a ser utilizada como parâmetro ao executar as
funções do pacote (character)
\item \strong{qtde_numeros_sorteio} Quantidade de números sorteados (numeric)
}
}
\examples{
# Retorna uma tibble contendo os produtos disponíveis
produtos_existentes()
}
