% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultados.R
\name{resultado_concurso}
\alias{resultado_concurso}
\title{Resultado de um Concurso Específico}
\usage{
resultado_concurso(produto = "megasena", concurso)
}
\arguments{
\item{produto}{Nome do jogo das Loterias que deseja atualizar a base.
Produtos disponíveis:
\itemize{
\item Megasena -> digite 'megasena' \link{megasena}
\item Lotofácil -> digite 'lotofacil' \link{lotofacil}
\item Quina -> digite 'quina' \link{quina}
\item Lotomania -> digite 'lotomania' \link{lotomania}
\item Timemania -> digite 'timemania' \link{timemania}
\item Super Sete -> digite 'supersete' \link{supersete}
\item Dia de Sorte -> digite 'diadesorte' \link{diadesorte}
}}

\item{concurso}{Número do concurso que deseja obter o resultado.}
}
\value{
Uma \code{tibble} - a quantidade de variáveis depende do produto.
}
\description{
Resultado de um Concurso Específico
}
\details{
Utilize esta função para receber um conjunto de dados contendo o resultado
de um concurso específico.
\strong{Importante}: o pacote atualiza as bases diariamente, porém, após instalá-lo
é importante que faça a atualização manual quando desejar os resultados mais
recentes. Utilize a função \link{atualizar_base_resultados} para este fim.
}
\examples{
# Receber os resultados do concurso número 1000 da Megasena
# No caso da Megasena, o resultado será uma tibble com 8 variáveis
resultado_concurso(produto = 'megasena', concurso = 1000)
}
