% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultados.R
\name{resultado_concursos}
\alias{resultado_concursos}
\title{Resultado de um Range de Concursos}
\usage{
resultado_concursos(
  produto = "megasena",
  num_concurso_inicial,
  num_concurso_final
)
}
\arguments{
\item{produto}{Nome do jogo das Loterias que deseja atualizar a base. Caso não
informado, por padrão será a Megasena \link{megasena}.
Produtos disponíveis:
\itemize{
\item Megasena -> digite 'megasena' \link{megasena}
\item Lotofácil -> digite 'lotofacil' \link{lotofacil}
\item Quina -> digite 'quina' \link{quina}
\item Lotomania -> digite 'lotomania' \link{lotomania}
\item Timemania -> digite 'timemania' \link{timemania}
\item Super Sete -> digite 'supersete' \link{supersete}
\item Dia de Sorte -> digite 'diadesorte' \link{diadesorte}
}}

\item{num_concurso_inicial}{Número do concurso inicial que deseja obter o resultado.}

\item{num_concurso_final}{Número do concurso final que deseja obter o resultado.
Este não é um parâmetro obrigatório. Caso informar apenas o concurso inicial,
serão retornados os dados desde o concurso inicial informado até o concurso mais
recente que existir no dataset.}
}
\value{
Uma \code{tibble} - a quantidade de variáveis depende do produto.
}
\description{
Utilize esta função para receber um conjunto de dados contendo os resultados
de um range de concursos, especificando o inicial e o final.
}
\details{
\strong{Importante}: o pacote atualiza as bases diariamente no repositório do GitHub,
porém, sua base local permanece com os dados do momento em que o pacote foi
instalado. Caso tenha intenção de usar as bases diretamente, sem chamar as
funções, reinstale o pacote para obter as bases mais recentes.
}
\examples{
# Visualizar concurso, números sorteados e se houve ganhador referentes aos
# resultados dos concursos entre 1000 a 1500 da Megasena
resultado_concursos(
   produto = 'megasena',
   num_concurso_inicial = 1000,
   num_concurso_final = 1500
 ) |> dplyr::select(concurso, numeros_sorteados, houve_ganhador)
}
