% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultados.R
\name{resultado_range_concursos}
\alias{resultado_range_concursos}
\title{Resultado de um Range de Concursos}
\usage{
resultado_range_concursos(
  produto = "megasena",
  num_concurso_inicial,
  num_concurso_final
)
}
\arguments{
\item{produto}{Nome do jogo das Loterias que deseja atualizar a base. Caso não
informado, por padrão será a Megasena \link{megasena}.
Produtos disponíveis:
\itemize{
\item Megasena -> digite 'megasena' \link{megasena}
\item Lotofácil -> digite 'lotofacil' \link{lotofacil}
\item Quina -> digite 'quina' \link{quina}
\item Lotomania -> digite 'lotomania' \link{lotomania}
\item Timemania -> digite 'timemania' \link{timemania}
\item Super Sete -> digite 'supersete' \link{supersete}
\item Dia de Sorte -> digite 'diadesorte' \link{diadesorte}
}}

\item{num_concurso_inicial}{Número do concurso inicial que deseja obter o resultado.}

\item{num_concurso_final}{Número do concurso final que deseja obter o resultado.
Este não é um parâmetro obrigatório. Caso informar apenas o concurso inicial,
serão retornados os dados desde o concurso inicial informado até o concurso mais
recente que existir no dataset.}
}
\value{
Uma \code{tibble} - a quantidade de variáveis depende do produto.
}
\description{
Utilize esta função para receber um conjunto de dados contendo os resultados
de um range de concursos, especificando o inicial e o final.
}
\details{
\strong{Importante}: o pacote atualiza as bases diariamente, porém, após instalá-lo
é importante que faça a atualização manual quando desejar os resultados mais
recentes. Utilize a função \link{atualizar_base_resultados} para este fim.
}
\examples{
# Visualizar a data, concurso e números sorteados referentes aos resultados
# dos concursos entre 1000 a 1500 da Megasena
# No caso da Megasena, o resultado será uma tibble com 8 variáveis
resultado_range_concursos(
   produto = 'megasena',
   num_concurso_inicial = 1000,
   num_concurso_final = 1500
 ) |> dplyr::select(data_apuracao, concurso, numeros_sorteados)
}
