% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultados.R
\name{resultado_todos}
\alias{resultado_todos}
\title{Resultados de Todos os Concursos}
\usage{
resultado_todos(produto = "megasena")
}
\arguments{
\item{produto}{Nome do jogo das Loterias que deseja atualizar a base. Caso não
informado, por padrão será a Megasena \link{megasena}.
Produtos disponíveis:
\itemize{
\item Megasena -> digite 'megasena' \link{megasena}
\item Lotofácil -> digite 'lotofacil' \link{lotofacil}
\item Quina -> digite 'quina' \link{quina}
\item Lotomania -> digite 'lotomania' \link{lotomania}
\item Timemania -> digite 'timemania' \link{timemania}
\item Super Sete -> digite 'supersete' \link{supersete}
\item Dia de Sorte -> digite 'diadesorte' \link{diadesorte}
}}
}
\value{
Uma \code{tibble} - a quantidade de variáveis depende do produto.
}
\description{
Utilize esta função para receber um conjunto de dados contendo todos os
resultados do produto escolhido.
}
\details{
\strong{Importante}: o pacote atualiza as bases diariamente no repositório do GitHub,
porém, sua base local permanece com os dados do momento em que o pacote foi
instalado. Caso tenha intenção de usar as bases diretamente, sem chamar as
funções, reinstale o pacote para obter as bases mais recentes.
}
\examples{
# Visualizar concurso, números sorteados e se houve ganhador referente aos
# resultados de todos os concursos da Megasena
resultado_todos(produto = 'megasena') |>
   dplyr::select(concurso, numeros_sorteados, houve_ganhador)
}
