% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultados.R
\name{resultados_todos}
\alias{resultados_todos}
\title{Todos os Resultados}
\usage{
resultados_todos(produto = "megasena")
}
\arguments{
\item{produto}{Nome do jogo das Loterias que deseja atualizar a base. Caso não
informado, por padrão será a Megasena \link{megasena}.
Produtos disponíveis:
\itemize{
\item Megasena -> digite 'megasena' \link{megasena}
\item Lotofácil -> digite 'lotofacil' \link{lotofacil}
\item Quina -> digite 'quina' \link{quina}
\item Lotomania -> digite 'lotomania' \link{lotomania}
\item Timemania -> digite 'timemania' \link{timemania}
\item Super Sete -> digite 'supersete' \link{supersete}
\item Dia de Sorte -> digite 'diadesorte' \link{diadesorte}
}}
}
\value{
Uma \code{tibble} - a quantidade de variáveis depende do produto.
}
\description{
Utilize esta função para receber um conjunto de dados contendo todos os
resultados do produto escolhido.
}
\details{
\strong{Importante}: o pacote atualiza as bases diariamente, porém, após instalá-lo
é importante que faça a atualização manual quando desejar os resultados mais
recentes. Utilize a função \link{atualizar_base_resultados} para este fim.
}
\examples{
# Visualizar a data, concurso e números sorteados referente aos resultados
# de todos os concursos da Megasena
# No caso da Megasena, o resultado será uma tibble com 8 variáveis
resultados_todos(produto = 'megasena') |>
   dplyr::select(data_apuracao, concurso, numeros_sorteados)
}
