% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultados.R
\name{resultados_ultimo_concurso}
\alias{resultados_ultimo_concurso}
\title{Resultado do Último Concurso}
\usage{
resultados_ultimo_concurso(produto = "megasena")
}
\arguments{
\item{produto}{Nome do jogo das Loterias que deseja atualizar a base. Caso não
informado, por padrão será a Megasena \link{megasena}.
Produtos disponíveis:
\itemize{
\item Megasena -> digite 'megasena' \link{megasena}
\item Lotofácil -> digite 'lotofacil' \link{lotofacil}
\item Quina -> digite 'quina' \link{quina}
\item Lotomania -> digite 'lotomania' \link{lotomania}
\item Timemania -> digite 'timemania' \link{timemania}
\item Super Sete -> digite 'supersete' \link{supersete}
\item Dia de Sorte -> digite 'diadesorte' \link{diadesorte}
}}
}
\value{
Uma \code{tibble} - a quantidade de variáveis depende do produto.
}
\description{
Utilize esta função para receber um conjunto de dados contendo todos o
resultado do último concurso do produto escolhido.
}
\details{
\strong{Importante}: o pacote atualiza as bases diariamente, porém, após instalá-lo
é importante que faça a atualização manual quando desejar os resultados mais
recentes. Utilize a função \link{atualizar_base_resultados} para este fim.
No caso desta função, essa rotina citada de atualização de base é acionada
automaticamente para o produto especificado.
}
\examples{
# Visualizar a data, concurso e números sorteados referente ao resultado do
# concurso mais recente da Megasena
# No caso da Megasena, o resultado será uma tibble com 8 variáveis
resultados_ultimo_concurso(produto = 'megasena') |>
   dplyr::select(data_apuracao, concurso, numeros_sorteados)
}
