% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultados.R
\name{resultado_concurso}
\alias{resultado_concurso}
\title{Resultado de um Concurso Específico}
\usage{
resultado_concurso(produto = "megasena", num_concurso)
}
\arguments{
\item{produto}{Nome do jogo das Loterias que deseja atualizar a base. Caso não
informado, por padrão será a Megasena \link{megasena}.
Produtos disponíveis:
\itemize{
\item Megasena -> digite 'megasena' \link{megasena}
\item Lotofácil -> digite 'lotofacil' \link{lotofacil}
\item Quina -> digite 'quina' \link{quina}
\item Lotomania -> digite 'lotomania' \link{lotomania}
\item Timemania -> digite 'timemania' \link{timemania}
\item Super Sete -> digite 'supersete' \link{supersete}
\item Dia de Sorte -> digite 'diadesorte' \link{diadesorte}
}}

\item{num_concurso}{Número do concurso que deseja obter o resultado.}
}
\value{
Uma \code{tibble} - a quantidade de variáveis depende do produto.
}
\description{
Utilize esta função para receber um conjunto de dados contendo o resultado
de um concurso específico.
}
\details{
\strong{Importante}: o pacote atualiza as bases diariamente, porém, após instalá-lo
é importante que faça a atualização manual quando desejar os resultados mais
recentes. Utilize a função \link{atualizar_base_resultados} para este fim.
}
\examples{
# Visualizar concurso, números sorteados e se houve ganhador referente ao
# resultado do concurso 1000 da Megasena
resultado_concurso(produto = 'megasena', num_concurso = 1000) |>
   dplyr::select(concurso, numeros_sorteados, houve_ganhador)
}
