% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supersete}
\alias{supersete}
\title{Super Sete}
\format{
Uma tibble contendo 9 variávies.
\itemize{
\item \strong{data_apuracao} Data do sorteio (character)
\item \strong{num_concurso} Numero do concurso (numeric)
\item \strong{num_1} Dezena da primeira coluna (numeric)
\item \strong{num_2} Dezena da segunda coluna (numeric)
\item \strong{num_3} Dezena da terceira coluna (numeric)
\item \strong{num_4} Dezena da quarta coluna (numeric)
\item \strong{num_5} Dezena da quinta coluna (numeric)
\item \strong{num_6} Dezena da sexta coluna (numeric)
\item \strong{num_7} Dezena da sétima coluna (numeric)
}
}
\source{
Site da Loteria Federal \url{https://loterias.caixa.gov.br/Paginas/Super-Sete.aspx}
}
\usage{
supersete
}
\description{
Conjunto de dados contendo todos os resultados do produto das Loterias
chamado Super Sete.
}
\note{
Para este produto a ordem dos números é a ordem do sorteio.
Para ganhar, é necessário acertar o número de cada coluna.
}
\keyword{datasets}
